Adding metadata and proper gating
