+++ .env
@@ -1,3 +1,9 @@
-NEXT_PUBLIC_PHOTOS_HOST=https://imagedelivery.net/dqFoxiedZNoncKJ9uqxz0g
-PHOTOS_HOST=https://imagedelivery.net/dqFoxiedZNoncKJ9uqxz0g
-PHOTO_API_HOST=https://image-test-importer.jsconf.cl
+NEXT_PUBLIC_SANITY_PROJECT_ID=t2zgeg0i
+NEXT_PUBLIC_SANITY_DATASET=dev
+NEXT_PUBLIC_SANITY_API_VERSION=
+NEXT_PUBLIC_SANITY_USE_CDN=True
+NEXT_PUBLIC_BASE_URL=https://fforres-google-auth.2023-images.pages.dev
+# CLERK
+NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY='pk_test_ZnVua3ktZ3JpZmZvbi04NC5jbGVyay5hY2NvdW50cy5kZXYk'
+CLERK_SECRET_KEY=''
+NEXT_PUBLIC_JSCL_API_URL="https://graphql-api.jsconfcl.workers.dev/graphql"
