+++ src/features/import/index.tsx
@@ -1,59 +1,45 @@
 "use client";
-import { ApolloClient, ApolloProvider, InMemoryCache } from "@apollo/client";
 import { Importer } from "./importer";
 import { Suspense, useEffect, useState } from "react";
 import { ErrorBoundary } from "./errorBoundary";
 import { useRouter } from "next/navigation";
-import { useAuth } from "@clerk/clerk-react";
+import { SignedIn, useAuth } from "@clerk/clerk-react";
 import { AllEventsQuery } from "../../gql/graphql";
+import { useIsSuperAdminSuspenseQuery } from "../../gql/jschileAPI";
 
-const useGetClient = (token: string) => {
-  return new ApolloClient({
-    uri: process.env.NEXT_PUBLIC_JSCL_API_URL,
-    cache: new InMemoryCache(),
-    headers: {
-      authorization: `Bearer ${token}`,
-    },
-  });
+const ImportImpl = ({
+  communityEvents,
+}: {
+  communityEvents: AllEventsQuery["allEventInstance"];
+}) => {
+  const router = useRouter();
+  const superAdminQuery = useIsSuperAdminSuspenseQuery();
+  const superAdmin = superAdminQuery.data.me.isSuperAdmin;
+  if (!superAdmin) {
+    router.push("/");
+    return null;
+  }
+
+  return (
+    <Suspense fallback={<div>Loading...</div>}>
+      <ErrorBoundary
+        onError={() => {
+          router.push("/", {});
+        }}
+      >
+        <Importer communityEvents={communityEvents} />
+      </ErrorBoundary>
+    </Suspense>
+  );
 };
 export const Import = ({
   communityEvents,
 }: {
   communityEvents: AllEventsQuery["allEventInstance"];
 }) => {
-  const { getToken } = useAuth();
-  const [token, setToken] = useState("");
-  const router = useRouter();
-  const client = useGetClient(token);
-  useEffect(() => {
-    const start = async () => {
-      const token = await getToken({
-        template: "API_AUTH",
-      });
-      if (!token) {
-        return;
-      }
-      setToken(token);
-    };
-    start().catch((e) => {
-      console.error(e);
-    });
-  }, []);
-
-  if (!token) {
-    return;
-  }
   return (
-    <ApolloProvider client={client}>
-      <Suspense fallback={<div>Loading...</div>}>
-        <ErrorBoundary
-          onError={() => {
-            router.push("/", {});
-          }}
-        >
-          <Importer communityEvents={communityEvents} />
-        </ErrorBoundary>
-      </Suspense>
-    </ApolloProvider>
+    <SignedIn>
+      <ImportImpl communityEvents={communityEvents} />
+    </SignedIn>
   );
 };
